Allow evaluation of branch name variable on zsh
The differences between bash and zsh's prompt evaluation apparently results in one fewer level of evaluation under zsh than bash, meaning that the security fix in magicmonty/bash-git-prompt#313 results in the branch name not bein rendered on zsh.
This just disables the fix, since zsh does not seem to be vulnerable to the same issue.